+++ Changelog.md
@@ -4,6 +4,7 @@
 
  - Added `appendInterpolation(_:format:)` and `appendInterpolation(_:if:)` string interpolation methods
  - `SetAlgebra` now has `+`, `+=`, `-`, and `-=` operators
+ - Added `~=` operator for `Error`
  - The package previously did not build on anything other than macOS. This has **hopefully** been fixed.
 
 ## 2.2.0